requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
flask-sock==0.7.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
# Production servers
gunicorn==21.2.0
gevent-websocket==0.12.1
